2004-10-22 03:45 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/source/rtl/spfiles.c     * changed hb_spFile() to always store file name with path in second       parameter (if given) even when it returns FALSE to eliminate       race condition in next calls to hb_spOpen()/hb_spCreate()       I'm still not sure about one thing - does Clipper check current       directory before SET_PATH if SET_DEFAULT is not set?       Anyhow the modification now will be done inside hb_spFile not       RDD subsystem.
  * xharbour/source/rdd/dbfcdx/dbfcdx1.c
    - removed all hacks with default paths from DBFCDX - now hb_spFile()
      is strong enough to make this job yourself ;-)

  * xharbour/source/pp/ppcore.c
    * casting for pointers bigger then int

+++ xharbour/source/pp/ppcore.c
@@ -1,5 +1,5 @@
 /*
- * $Id: ppcore.c,v 1.175 2004/10/20 22:40:50 ronpinkas Exp $
+ * $Id: ppcore.c,v 1.176 2004/10/21 17:04:41 ronpinkas Exp $
  */
 
 /*
@@ -2294,7 +2294,7 @@ int hb_pp_ParseExpression( char * sLine, char * sOutLine )
                  if( hb_comp_PPTrace )
                  {
                     fprintf( hb_comp_PPTrace, "%s(%i) >%.*s<\n", hb_comp_files.pLast->szFileName, hb_comp_iLine - 1, i, ptrb );
-                    fprintf( hb_comp_PPTrace, "#defined >%.*s<\n\n", ptri - ptrb, ptro );
+                    fprintf( hb_comp_PPTrace, "#defined >%.*s<\n\n", ( int ) ( ptri - ptrb ), ptro );
                  }
 
                  hb_pp_Stuff( ptro, ptrb, i, ptri - ptrb, lens + 1 );
